fix bug of max split size

+++ ros/src/util/packages/autoware_rviz_plugins/rosbag_controller/src/autoware_rosbag_plugin.h
@@ -69,7 +69,7 @@ public:
       std::string filename;
       std::vector<std::string> topics;
       int max_duration;
-      uint32_t max_size;
+      uint64_t max_size;
     } RecordParam;
 
 protected:
